refactor forum queries and mutations
