+++ src/components/Menus/Forum/Items/ForumPostActionsFavoriteItem.tsx
@@ -1,7 +1,7 @@
 import {Menu} from 'react-native-paper';
 import React from 'react';
 import {PostData} from '../../../../libraries/Structs/ControllerStructs';
-import {useForumPostBookmarkMutation} from '../../../Queries/Forum/ForumPostBookmarkQueries';
+import {useForumPostBookmarkMutation} from '../../../Queries/Forum/ForumPostBookmarkMutations';
 import {useTwitarr} from '../../../Context/Contexts/TwitarrContext';
 import {ForumPostListActions} from '../../../Reducers/Forum/ForumPostListReducer';
 import {StateLoadingIcon} from '../../../Icons/StateLoadingIcon';
