+++ src/components/Queries/Forum/ForumThreadRelationQueries.tsx
@@ -1,28 +1,5 @@
-import {useTokenAuthMutation} from '../TokenAuthMutation';
-import axios, {AxiosResponse} from 'axios/index';
 import {useTokenAuthPaginationQuery} from '../TokenAuthQuery';
 import {ForumSearchData} from '../../../libraries/Structs/ControllerStructs';
-
-interface ForumRelationCreateProps {
-  forumID: string;
-  relationType: 'favorite' | 'mute';
-  action: 'create' | 'delete';
-}
-
-const relationQueryHandler = async ({
-  forumID,
-  relationType,
-  action,
-}: ForumRelationCreateProps): Promise<AxiosResponse<void>> => {
-  if (action === 'delete') {
-    return await axios.delete(`/forum/${forumID}/${relationType}`);
-  }
-  return await axios.post(`/forum/${forumID}/${relationType}`);
-};
-
-export const useForumRelationMutation = () => {
-  return useTokenAuthMutation(relationQueryHandler);
-};
 
 export interface ForumRelationQueryParams {
   start?: number;
