+++ src/components/Screens/Forum/Thread/ForumThreadMutesScreen.tsx
@@ -8,7 +8,7 @@ import {HeaderButtons} from 'react-navigation-header-buttons';
 import {ForumThreadScreenSortMenu} from '../../../Menus/Forum/ForumThreadScreenSortMenu';
 import {ForumThreadsRelationsView} from '../../../Views/Forum/ForumThreadsRelationsView';
 import {ForumFilter} from '../../../../libraries/Enums/ForumSortFilter';
-import {ForumRelationQueryType} from '../../../Queries/Forum/ForumRelationQueries';
+import {ForumRelationQueryType} from '../../../Queries/Forum/ForumThreadRelationQueries';
 
 export type Props = NativeStackScreenProps<
   ForumStackParamList,
