+++ src/components/Screens/Forum/Thread/ForumThreadRecentScreen.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import {ForumRelationQueryType} from '../../../Queries/Forum/ForumRelationQueries';
+import {ForumRelationQueryType} from '../../../Queries/Forum/ForumThreadRelationQueries';
 import {ForumThreadsRelationsView} from '../../../Views/Forum/ForumThreadsRelationsView';
 
 export const ForumThreadRecentScreen = () => {
