+++ src/components/Screens/Forum/Thread/ForumThreadUserScreen.tsx
@@ -7,7 +7,7 @@ import {RefreshControl} from 'react-native';
 import {PaddedContentView} from '../../../Views/Content/PaddedContentView';
 import {Text} from 'react-native-paper';
 import {ForumThreadFlatList} from '../../../Lists/Forums/ForumThreadFlatList';
-import {useForumSearchQuery} from '../../../Queries/Forum/ForumSearchQueries';
+import {useForumSearchQuery} from '../../../Queries/Forum/ForumThreadSearchQueries';
 import {ListTitleView} from '../../../Views/ListTitleView';
 import {getUserBylineString} from '../../../Text/Tags/UserBylineTag';
 import {CommonStackComponents, CommonStackParamList} from '../../../Navigation/CommonScreens';
